# backend/api/openapi.yaml
openapi: 3.1.0
info:
  title: KatalogitAI API
  version: 1.0.0
  description: >
    Backend for KatalogitAI — AI product photoshoots for fashion vendors.
    All endpoints (except webhooks) require `Authorization: Bearer <Firebase ID token>`.
    The credit ledger lives in Cloud SQL; catalogue lives in Firestore.
servers:
  - url: /api/v1
    description: Same-origin via Firebase Hosting rewrite to Cloud Run
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiResponse:
      type: object
      properties:
        ok: { type: boolean, example: true }
        data: { type: object }
      required: [ok]
...
    OrderCreateRequest:
      type: object
      required: [packId]
      properties:
        packId: { type: string, example: pack_100 }
    Order:
      type: object
      properties:
        orderId: { type: string }
        credits: { type: number }
        amountPaise: { type: integer }
        currency: { type: string, example: INR }
        status: { type: string, enum: [created, paid, failed, refunded] }
...
  responses:
    Ok:
      description: Success
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ApiResponse'
              - type: object
                properties:
                  data: { type: object }
...
security:
  - bearerAuth: []
paths:
...
  /stores/{storeId}/credits/ledger:
    get:
      summary: Credit transaction history
      parameters:
        - { $in: query, name: cursor, schema: { type: string } }
        - { $in: query, name: limit, schema: { type: integer, default: 25, maximum: 100 } }
      responses:
        '200': { $ref: '#/components/responses/Ok' }
...
  /credits/orders:
    post:
      summary: Create a Razorpay top-up order
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/OrderCreateRequest' }
      responses:
        '201':
          description: Razorpay order created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/Order' }
...
  /admin/credits:
    post:
      summary: Admin — adjust credits (requires role=admin claim)
      security:
        - bearerAuth: []
      responses:
        '200': { $ref: '#/components/responses/Ok' }